Simplify ProtocolMetrics. Move files.
